scripts/pyproject.toml
[project]
name = "technical-docs-scripts"
version = "0.1.0"
description = "Scripts to extract technical documentation for RAG evaluation"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "tqdm>=4.66.0",
    "pypandoc-binary>=1.13",
    "pyyaml>=6.0",
]

[dependency-groups]
dev = [
    # Code Quality
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "bandit[toml]>=1.7.5",
    # Type stubs
    "types-tqdm>=4.66.0",
    "types-pyyaml>=6.0",
]

[tool.ruff]
line-length = 88
target-version = "py311"
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "venv",
    ".pytest_cache",
    ".mypy_cache",
    ".ruff_cache",
    "data",
]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "ARG",  # flake8-unused-arguments
    "SIM",  # flake8-simplify
    "PTH",  # flake8-use-pathlib
    "RUF",  # ruff-specific rules
    "N",    # pep8-naming
    "D",    # pydocstyle
    "ANN",  # flake8-annotations
    "S",    # flake8-bandit (security)
    "C90",  # mccabe complexity
    "PLC0415",  # import-outside-toplevel - forbids inline imports
]
ignore = [
    "D100",  # Missing docstring in public module (we have module docstrings)
    "D104",  # Missing docstring in public package
    "D107",  # Missing docstring in __init__
]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "ANN", "S101"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
no_implicit_optional = true
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.bandit]
exclude_dirs = ["/tests", "/.venv", "/data"]
skips = ["B101"]
